moved the hello and identity logic to the commands module
