+++ src/core/commands/mod.rs
@@ -1,1 +1,2 @@
-pub mod cold_start;
+pub mod hello;
+pub mod identify;
