--- src/core/events/handlers/mod.rs
@@ -1,1 +1,0 @@
-pub mod hello;
